Purge legacy x86/32-bit type aliases from IXLandLinuxRuntime consumers
Removed all uses of sqword_t, qword_t, dword_t, sdword_t, word_t, byte_t, uint_t, int_t across 20 files in IXLandLinuxRuntime, replacing with explicit fixed-width C integer types (uint8_t, uint32_t, uint64_t, int64_t, uintptr_t, addr_t).

Files changed:
- fs/fd.c: sys_close, sys_dup, sys_dup2, sys_dup3, sys_fcntl, sys_fcntl32
- fs/mount.c: sys_mount, sys_umount2
- fs/sock.c: sys_sendmsg, sys_recvmsg, sys_accept4, sys_recvmmsg, sys_sendmmsg, sys_socketcall
- kernel/aarch64/calls.c: a64_syscall_name
- kernel/calls.h: verified declarations already correct
- kernel/epoll.c: sys_epoll_create, sys_epoll_ctl, sys_epoll_wait, sys_epoll_pwait
- kernel/eventfd.c: sys_eventfd, sys_eventfd2
- kernel/exit.c: sys_waitid
- kernel/fork.c: sys_clone, sys_fork, sys_vfork
- kernel/fs.c: 60+ functions using dword_t/int_t for syscall params
- kernel/futex.c: futex_load, futex_wait, futex_wake, sys_futex, sys_set_robust_list, sys_get_robust_list
- kernel/getset.c: sys_getuid/sys_geteuid/sys_getgid/sys_getegid variants, sys_setuid/sys_setgid, sys_setresuid/sys_setresgid, sys_getresuid/sys_getresgid, sys_getgroups/sys_setgroups, sys_capget/sys_capset, sys_personality
- kernel/group.c: sys_setpgid, sys_setpgrp, sys_setsid, sys_getsid
- kernel/ipc.c: sys_ipc
- kernel/log.c: sys_syslog
- kernel/misc.c: sys_prctl, sys_arch_prctl, sys_reboot
- kernel/mmap.c: do_mmap, mmap_common, sys_mmap, sys_mmap2, sys_munmap, sys_mremap, sys_mprotect, sys_madvise, sys_mbind, sys_mlock, sys_msync
- kernel/poll.c: sys_select, sys_poll, sys_pselect, sys_ppoll
- tcti/aarch64/gadgets_memory.c: memory probe
- Tests/IXLandLinuxRuntimeSystemTests/Guest/ELFLoader/GuestELFLoaderTests.m: byte_t → uint8_t

Build: TEST BUILD SUCCEEDED (clean build confirmed)

+++ Sources/IXLandLinuxRuntime/kernel/calls.h
@@ -90,7 +90,7 @@ uint32_t sys_close(fd_t fd);
 uint32_t sys_fsync(fd_t f);
 uint32_t sys_flock(fd_t fd, uint32_t operation);
 int64_t sys_pipe(addr_t pipe_addr);
-int64_t sys_pipe2(addr_t pipe_addr, int64_t flags);
+int64_t sys_pipe2(addr_t pipe_addr, int32_t flags);
 struct pollfd_ {
     fd_t fd;
     uint16_t events;
